Move message to separate file
